requirements.txt
Django==2.0.3
gunicorn==20.0.4
cryptography==2.1.4
pycrypto==2.6.1
asn1crypto==0.24.0
django-heroku==0.3.1
